+++ src/nodejs/lambda-layers/database-util/src/query/submission.js
@@ -5,7 +5,7 @@ const workflow = require('./workflow.js');
 const table = 'submission';
 // const allFields = ['id', 'name', 'user_id', 'daac_id', 'conversation_id', 'workflow_id', 'workflow_name', 'step_name', 'status', 'forms', 'action_data', 'form_data', 'metadata', 'created_at', 'last_change', 'lock'];
 
-const allFields = ['id', 'name', 'initiator', 'workflow_id', 'hidden', 'conversation_id', 'workflow_name', 'daac_id', 'daac_name', 'step_data', 'step_status_label', 'step_name', 'status', 'forms', 'action_data', 'form_data', 'metadata', 'created_at', 'last_change', 'lock', 'contributor_ids', 'copy', 'origin_id'];
+const allFields = ['id', 'name', 'initiator', 'workflow_id', 'hidden', 'conversation_id', 'workflow_name', 'daac_id', 'daac_name', 'code', 'assigned_daacs', 'step_data', 'step_status_label', 'step_name', 'status', 'forms', 'action_data', 'form_data', 'metadata', 'created_at', 'last_change', 'lock', 'contributor_ids', 'copy', 'origin_id'];
 const customFields = ['id', 'name', 'data_producer_name', 'initiator', 'workflow_id', 'hidden', 'conversation_id', 'workflow_name', 'daac_id', 'daac_name', 'step_data', 'step_status_label', 'step_name', 'status', 'created_at', 'last_change', 'lock', 'contributor_ids', 'copy', 'origin_id'];
 
 const fieldMap = {
@@ -33,7 +33,9 @@ const fieldMap = {
   lock: '(EXISTS(SELECT edpuser_id FROM submission_lock WHERE submission_lock.id = submission.id)) "lock"',
   contributor_ids: 'submission.contributor_ids',
   copy: '(EXISTS(SELECT edpuser_id FROM submission_copy WHERE submission_copy.id = submission.id)) "copy"',
-  origin_id: 'submission_copy.origin_id'
+  origin_id: 'submission_copy.origin_id',
+  code: 'publication_accession_association.code',
+  assigned_daacs: 'submission_daac_assignment.assigned_daacs'
 };
 const refs = {
   initiator_ref: {
@@ -110,7 +112,8 @@ const refs = {
             ['form_id', 'step.form_id'],
             ['service_id', 'step.service_id'],
             ['data', 'step.data'],
-            ['daac_only', 'form.daac_only']
+            ['daac_only', 'form.daac_only'],
+            ['upload_step_id', 'upload_step.id']
           ],
           strip: true,
           alias: 'step_data'
@@ -118,11 +121,17 @@ const refs = {
       ],
       from: { 
         base: 'step', 
-        joins: [{
-          type: 'left_join',
-          src: 'form',
-          on: { left: 'step.form_id', right: 'form.id' }
-        }]
+        joins: [
+          {
+            type: 'left_join',
+            src: 'form',
+            on: { left: 'step.form_id', right: 'form.id' }
+          }, {
+            type: 'left_join',
+            src: 'upload_step',
+            on: { left: 'step.step_name', right: 'upload_step.step_name' }
+          }
+        ]
       },
       alias: 'step'
     }
@@ -141,6 +150,35 @@ const refs = {
     type: 'left_join',
     src: 'submission_form_data_pool',
     on:{ left: 'submission_form_data_pool.id', right: 'submission.id'}
+  },
+  publication_accession_association: {
+    type: 'left_join',
+    src: 'publication_accession_association',
+    on:{ left: 'submission.id', right: 'publication_accession_association.publication_submission_id'}
+  },
+  code : {
+    type: 'natural_left_join',
+    src: {
+      type: 'select',
+      fields: [`code.submission_id, COALESCE(JSONB_AGG(JSONB_STRIP_NULLS(JSONB_BUILD_OBJECT('code', code.code, 'daac_id', code.daac_id, 'daac_name', daac.short_name))), \'{}\'::JSONB) assigned_daacs`],
+      from: { 
+        base: 'code',
+        joins: [{
+          type: 'left_join',
+          src: 'daac',
+          on: { left: 'code.daac_id', right: 'daac.id' }
+        }]
+      
+      },
+      where: {
+        filters: [
+          { field: 'code.submission_id', param: 'id' }
+        ]
+      },
+      group: 'code.submission_id',
+      alias: 'submission_daac_assignment'
+    },
+    on:{ left: 'submission.id', right: 'submission_daac_assignment.submission_id'}
   }
 };
 
@@ -280,7 +318,7 @@ const findById = (params) => sql.select({
   fields: fields(allFields),
   from: {
     base: table,
-    joins: [refs.submission_status, refs.initiator_ref, refs.submission_metadata, refs.submission_action_data, submission_form_data(params.privileged_user), refs.step, refs.workflow, refs.daac, refs.submission_form_data_pool, refs.submission_copy, submissionPrivilegedUsers()]
+    joins: [refs.submission_status, refs.code, refs.publication_accession_association, refs.initiator_ref, refs.submission_metadata, refs.submission_action_data, submission_form_data(params.privileged_user), refs.step, refs.workflow, refs.daac, refs.submission_form_data_pool, refs.submission_copy, submissionPrivilegedUsers()]
   },
   where: {
     filters: [
@@ -425,9 +463,13 @@ FROM submission
 WHERE submission.id = {{id}}`;
 
 const initialize = (params) => `
-INSERT INTO submission(initiator_edpuser_id${params.daac_id ? ', daac_id' : ''}${params.name ? ', name' : ''}, contributor_ids)
-VALUES ({{user_id}}${params.daac_id ? ', {{daac_id}}' : ''}${params.name ? ', {{name}}' : ''}, ARRAY[{{user_id}}::UUID])
-RETURNING *`;
+WITH new_submission AS (
+  INSERT INTO submission(initiator_edpuser_id${params.daac_id ? ', daac_id' : ''}${params.name ? ', name' : ''}, contributor_ids)
+  VALUES ({{user_id}}${params.daac_id ? ', {{daac_id}}' : ''}${params.name ? ', {{name}}' : ''}, ARRAY[{{user_id}}::UUID])
+  RETURNING *
+)
+${params.code ? ', association AS (INSERT INTO publication_accession_association SELECT new_submission.id, {{accession_submission_id}}, {{code}} FROM new_submission)' : ''}
+SELECT * FROM new_submission`;
 
 const updateName = () => `
 UPDATE submission
@@ -699,11 +741,11 @@ const getStepName = () => `
 SELECT step_name FROM submission_status WHERE id = {{id}}
 `;
 
-// TODO - Upate this query's complexity and to use sql builder
+// TODO - Update this query's complexity and to use sql builder
 const getSubmissionDetailsById = (params) => `
-WITH step_visibility AS (SELECT step.*, form.daac_only FROM step LEFT JOIN form ON step.form_id = form.id),
+WITH step_visibility AS (SELECT step.*, upload_step.id AS upload_step_id, form.daac_only FROM step LEFT JOIN form ON step.form_id = form.id LEFT JOIN upload_step ON step.step_name = upload_step.step_name),
 filteredForm AS (SELECT * FROM form ${params.privilegedUser === false ? `WHERE form.daac_only=false`: ``})
-SELECT submission.id id, conversation_id, submission.created_at created_at, daac.long_name daac_name, 
+SELECT submission.id id, conversation_id, submission.created_at created_at,
 submission.hidden hidden,
 JSONB_BUILD_OBJECT('name', edpuser1.name, 'id', edpuser1.id) initiator,
 submission_status.last_change last_change,
@@ -712,7 +754,7 @@ submission_form_data_pool.data::json->'data_product_name_value' data_product_nam
 array_agg(DISTINCT JSONB_BUILD_OBJECT('id', edpuser2.id, 'name', edpuser2.name)) as contributors,
 JSONB_BUILD_OBJECT('id', workflow.id, 'name', workflow.long_name, 'steps',
 array_agg(DISTINCT step_edge.step_name)) workflow,
-JSONB_STRIP_NULLS(JSONB_BUILD_OBJECT('type', step_visibility.type, 'name', step_visibility.step_name,'action_id', step_visibility.action_id, 
+JSONB_STRIP_NULLS(JSONB_BUILD_OBJECT('type', step_visibility.type, 'name', step_visibility.step_name,'action_id', step_visibility.action_id, 'upload_step_id', step_visibility.upload_step_id, 
 'form_id', step_visibility.form_id,'service_id', step_visibility.service_id, 'data', step_visibility.data, 'daac_only', step_visibility.daac_only)) step_data,
 CASE 
   WHEN filteredForm.id is null THEN '[]'
@@ -723,8 +765,6 @@ CASE
       'long_name', filteredForm.long_name, 'daac_only', filteredForm.daac_only, 'submitted_at', submission_form_data.submitted_at))) 
 END forms, submission_metadata.metadata metadata
 FROM submission
-JOIN daac
-ON submission.daac_id = daac.id
 JOIN edpuser edpuser1
 ON submission.initiator_edpuser_id = edpuser1.id
 JOIN submission_status
@@ -746,9 +786,9 @@ ON submission_form_data.form_id = filteredForm.id
 JOIN submission_metadata
 ON submission.id = submission_metadata.id
 WHERE submission.id= {{id}}
-GROUP BY submission.id, daac.long_name, edpuser1.name, edpuser1.id,
+GROUP BY submission.id, edpuser1.name, edpuser1.id,
 submission_status.last_change, workflow.long_name, workflow.id,
-submission_form_data_pool.data, step_visibility.type, step_visibility.step_name, step_visibility.action_id,
+submission_form_data_pool.data, step_visibility.type, step_visibility.step_name, step_visibility.action_id, step_visibility.upload_step_id,
 step_visibility.form_id, step_visibility.service_id, step_visibility.data, step_visibility.daac_only, filteredForm.id, 
 filteredForm.long_name, submission_form_data.submitted_at, submission_metadata.metadata;
 `;
@@ -850,7 +890,14 @@ WHERE code = {{code}}
 const createCode = () => `
 INSERT INTO code(submission_id, daac_id)
 Values({{submissionId}}, {{daacID}})
+ON CONFLICT DO NOTHING
 RETURNING *
+`
+
+const deleteCodes = (params) => `
+DELETE FROM code 
+WHERE submission_id={{submissionId}} 
+AND daac_id=ANY(ARRAY[${params.daacs.map(id => `'${id}'`).join(',')}]::UUID[])
 `
 
 module.exports.findAll = findAll;
@@ -901,3 +948,4 @@ module.exports.updateStatusStepReviewApproval = updateStatusStepReviewApproval;
 module.exports.stepCleanup = stepCleanup;
 module.exports.checkCode = checkCode;
 module.exports.createCode = createCode;
+module.exports.deleteCodes = deleteCodes;