+++ src/nodejs/lambda-layers/database-util/src/query/upload.js
@@ -1,0 +1,7 @@
+const sql = require('./sql-builder.js');
+
+const findAllUploadSteps = () => 'SELECT upload_step.* FROM upload_step';
+const findUploadStepById = () => `${findAllUploadSteps()} WHERE upload_step.id = {{id}}`;
+
+module.exports.findAllUploadSteps = findAllUploadSteps;
+module.exports.findUploadStepById  = findUploadStepById;