+++ src/nodejs/lambda-layers/message-util/src/create-email.js
@@ -3,6 +3,7 @@ const { getNewSubmissionTemplate } = require('./templates/new-submission');
 const { getNewSubmissionDAACTemplate } = require('./templates/new-submission-daac');
 const { getDMTemplate } = require('./templates/direct-message');
 const { getReviewerAddedTemplate } = require('./templates/review-template');
+const { getAssignedDaacCodeTemplate } = require('./templates/assigned-daac-codes.js');
 
 const envUrl = process.env.ROOT_URL;
 
@@ -17,6 +18,8 @@ const createEmailHtml = async (params) => {
       ? getNewSubmissionTemplate(params, envUrl) : getNewSubmissionDAACTemplate(params, envUrl));
   } if (params.eventMessage.event_type === 'review_required') {
     return getReviewerAddedTemplate(params, envUrl);
+  } if (params.eventMessage.event_type === 'daac_assignment') {
+    return getAssignedDaacCodeTemplate(params, envUrl);
   }
   return getDefaultStepPromotion(params, envUrl);
 };